Implement what should work for evaluate - Fixes #1
This version will not work if example has an `(attribute, value)` pair where
`attribute` is used in the tree, but `value` is not.

If this is necessary, it should be handled underneath the for loop. If
control flow gets there, no branch of the tree is suitable. What should be
done is go over all leaves of the tree with root `node`, and return the
most common classification.