Add print_node
This prints the entire tree whose root is the node on which it is called.

This should (probably) not make it into the hand-in version of the code,
as it isn't really useful for the task at hand... But, I think it could be
useful in debugging, to inspect the tree given/generated by a function.

The format in which this outputs is... Well, kinda disgusting. I'll adjust
it a bit in a second