Implement a basic pruner
This pruner will only prune the one node whose removal best improves the
performance of the tree. Not really what I want the final version to do...
But a start. Perhaps this can be a helper, to be called until performance
improvement is below some epsilon?

Also, add a function to node: root.remove_descendant(n) will search the entire
tree (excepting the root) for n, and remove it if its found. It returns
whether a node was removed.