Implement a brute force pruner
This prunes one node at a time (greedy local search) until the difference
in accuracy between successive runs falls below a hardcoded epsilon value.

For now, set it to 0. It can be tuned further later

This... Isn't exactly very performant, and it gets caught in local maxima
of accuracy, but it is a pruner. I think...